Added a function to get the pseudo root
The pseudo root is made up term.
As we are only trying to minimize crossings,
there is no need to consider graphs as directed.
If we treat them as undirected and can get the undirected trees,
this allows as to use any node as a root.
However, given that the actual tree is directed,
using any node as a root will result in a bad tree (but no crossings)
The pseudo root is the node such that,
if we reconsidered that the tree is directed,
would reach the maximal number of nodes.
Using the pseudo root as the root of the tree when displaying it
will result in most edges flowing in the same direction.